base_requirements.txt
apache-airflow==2.9.2
apache-airflow-providers-docker
numpy
pandas